refactor: split grand fleet role pipelines
