+++ extensions/grand-fleet/AGENTS.md
@@ -5,13 +5,13 @@
 ## Role
 
 환경변수 `PI_GRAND_FLEET_ROLE`에 따라 2가지 모드로 동작:
-- **admiralty**: 지휘소 모드 — JSON-RPC 서버 기동, 함대 관리, 시스템 프롬프트 전체 교체
-- **fleet**: 함대 모드 — JSON-RPC 클라이언트로 Admiralty에 접속, Grand Fleet Context 프롬프트 append
+- **admiralty**: 지휘소 모드 — JSON-RPC 서버 기동, 함대 관리, 시스템 프롬프트 전체 교체 (Fleet Admiral 페르소나)
+- **fleet**: 함대 모드 — JSON-RPC 클라이언트로 Admiralty에 접속, Grand Fleet Context 프롬프트 append (Admiral 제독 페르소나)
 - **미설정**: 아무 동작 없음 (기존 단일 함대 모드)
 
 ## 4-Tier Naval Hierarchy (4계층 해군 위계)
 
-Grand Fleet 도입과 함께 체계화된 4단계 위계 구조입니다:
+Grand Fleet 도입과 함께 체화된 4단계 위계 구조입니다:
 
 | Tier | Entity | Persona / Target | Role |
 |------|--------|-----------------|------|
@@ -29,17 +29,19 @@ Grand Fleet 도입과 함께 체계화된 4단계 위계 구조입니다:
 ## 2계층 아키텍처 (Runtime Layer)
 
 시스템의 실행 레이어 구조입니다 (위계 구조와 별개):
+
+### Connectivity Runtime Layer (연결 런타임)
 환경변수만으로 동작하는 범용 메커니즘. Formation Strategy와 무관하게 독립 동작.
 - IPC 서버/클라이언트 (ipc/)
 - 함대 레지스트리 (admiralty/fleet-registry.ts)
 - 시스템 프롬프트 분기 (prompts.ts)
-- 도구 등록 (admiralty/tools.ts)
+- 도구 등록 (admiralty/pi-tools.ts, fleet/pi-tools.ts)
 
 ### Deployment Runtime Layer (배치 런타임)
 Admiralty가 직접 Fleet를 파견할 때 사용하는 실행 레이어.
-- Admiralty 수명주기와 IPC 서버 일원화 (admiralty/register.ts)
+- Admiralty 수명주기와 IPC 서버 일원화 (admiralty/pi-events.ts, admiralty/runtime.ts)
 - tmux 통합 (formation/tmux.ts)
-- MCP deploy 도구 (admiralty/tools.ts)
+- MCP deploy 도구 (admiralty/pi-tools.ts)
 
 ## Domain Boundary Rules
 
@@ -50,9 +52,10 @@ grand-fleet/  →  core/ (임포트 허용)
                  ✗ fleet/carriers/
 ```
 
-- `grand-fleet/` → `core/` 임포트만 허용
-- `grand-fleet/` → `fleet/`, `fleet/admiral/`, `fleet/carriers/` 임포트 **금지**
-- 다른 extension → `grand-fleet/` 임포트 **금지**
+- `grand-fleet/` → `core/` 임포트만 허용.
+- `grand-fleet/` → `fleet/`, `fleet/admiral/`, `fleet/carriers/` 임포트 **절대 금지**.
+- `grand-fleet/`는 `extensions/fleet/**`의 기능을 사용하지 않고 독자적인 `prompts.ts` 조립 솔기(seam)와 `globalThis` 런타임 상태를 통해 동작한다.
+- 다른 extension → `grand-fleet/` 임포트 **금지**.
 
 ## Environment Variables
 
@@ -67,21 +70,30 @@ grand-fleet/  →  core/ (임포트 허용)
 
 | File | Role |
 |------|------|
-| `index.ts` | 엔트리포인트 — 환경변수 분기 + globalThis 초기화 |
-| `types.ts` | 공용 타입, 상수, globalThis 키 |
-| `prompts.ts` | Admiralty + Fleet 시스템 프롬프트 |
+| `index.ts` | 엔트리포인트 facade — 환경변수 분기 + globalThis 초기화 + 역할 dispatcher |
+| `types.ts` | 공용 타입, 상수, globalThis 키, 역할별 runtime state 타입 |
+| `prompts.ts` | Admiralty + Fleet 시스템 프롬프트 조립 SSOT (Local Composition Seam) |
 | `ipc/protocol.ts` | ndJSON 프레이밍 + 파싱 |
 | `ipc/server.ts` | Admiralty JSON-RPC 서버 |
 | `ipc/client.ts` | Fleet JSON-RPC 클라이언트 |
 | `ipc/methods.ts` | 메서드 핸들러 |
-| `admiralty/register.ts` | Admiralty 모드 와이어링 |
-| `admiralty/tools.ts` | grand_fleet_deploy / dispatch / broadcast / status 도구 |
+| `admiralty/register.ts` | Admiralty 역할 facade |
+| `admiralty/pi-events.ts` | Admiralty `before_agent_start`, `session_start`, `session_shutdown` owner |
+| `admiralty/runtime.ts` | Admiralty server, registry, presenter, roster listener disposer owner |
+| `admiralty/pi-tools.ts` | grand_fleet_deploy / dispatch / recall / broadcast / status 도구 |
 | `admiralty/fleet-registry.ts` | 함대 등록/해제/상태 관리 |
 | `admiralty/report-renderer.ts` | 함대 보고서 TUI 렌더링 |
 | `admiralty/status-overlay.ts` | Admiralty 상황판 오버레이 (Alt+G) |
+| `admiralty/status-overlay-keybind.ts` | Admiralty Alt+G keybind와 popup guard |
 | `admiralty/roster-widget.ts` | Admiralty 함대 로스터 위젯 |
-| `fleet/register.ts` | Fleet 모드 와이어링 + Carrier 상태 수집/전송 |
+| `fleet/register.ts` | Fleet 역할 facade |
+| `fleet/pi-events.ts` | Fleet `before_agent_start`, `session_start`, `message_end`, `agent_end`, `session_shutdown` owner |
+| `fleet/pi-commands.ts` | `fleet:grand-fleet:connect` / `disconnect` 커맨드 owner |
+| `fleet/pi-tools.ts` | `mission_report` 도구 owner |
+| `fleet/runtime.ts` | FleetClient, heartbeat/status timers, mission/report buffer owner |
+| `fleet/status-source.ts` | Carrier snapshot, fleet ping payload, overlay runtime snapshot 입력 owner |
 | `fleet/status-overlay.ts` | Fleet 상황판 오버레이 (Alt+G) |
+| `fleet/status-overlay-keybind.ts` | Fleet Alt+G keybind와 popup guard |
 | `fleet/reporter.ts` | 작전 보고 모듈 |
 | `overlay-frame.ts` | 오버레이 프레임 공용 유틸 (border, truncation) |
 | `text-sanitize.ts` | ANSI/제어문자 제거 유틸 |
@@ -89,13 +101,31 @@ grand-fleet/  →  core/ (임포트 허용)
 
 ## Core Rules
 
-- **부팅 제어** — `boot/` 확장의 `__fleet_boot_config__` globalThis 플래그로 로드 여부 결정. 역할 미설정 시 grand-fleet 전체가 비활성화된다.
-- **역할별 도구/커맨드 격리** — Admiralty 도구(`grand_fleet_deploy`, `dispatch`, `broadcast`, `status`)는 `admiralty` 역할에서만 등록. Fleet 모드에서는 `connect`/`disconnect` 커맨드만 등록.
-- **프롬프트 수준 격리** — Admiralty 모드에서 `before_agent_start`로 시스템 프롬프트를 전체 교체. Fleet 모드에서는 연결 상태일 때만 Grand Fleet Context를 append.
-- **하이브리드 접속** — Fleet 모드에서 `PI_GRAND_FLEET_SOCK` env var가 있으면 자동 접속, 없으면 `/fleet:grand-fleet:connect`로 수동 접속.
-- **Prompt text lives in `prompts.ts`** — AI 프롬프트는 `prompts.ts`에 분리.
-- **globalThis로 모듈 간 상태 공유** — 세션 전환에도 유지.
-- **`index.ts` is for wiring only** — 비즈니스 로직은 하위 모듈에.
-- **슬래시 커맨드**: `fleet:grand-fleet:<feature>` 형식.
-- **키바인딩** — `Alt+G`: Grand Fleet Status 오버레이. 역할에 따라 Admiralty/Fleet 뷰를 자동 분기.
-- **외부 입력 정화** — 오버레이/위젯에서 렌더링되는 외부 문자열(fleet ID, zone, mission objective 등)은 반드시 `stripControlChars()`로 ANSI/제어문자를 제거한 후 렌더링. Carrier task 텍스트는 `sanitizeTaskText()`로 민감 토큰 마스킹 + 80자 절단.
+### 1. 부팅 및 역할 제어
+- `boot/` 확장의 `__fleet_boot_config__` globalThis 플래그로 로드 여부 결정.
+- 환경변수 `PI_GRAND_FLEET_ROLE`이 미설정된 경우 grand-fleet 전체가 비활성화된다.
+- **index.ts**는 오직 wiring과 dispatcher 역할만 수행하며, 비즈니스 로직은 하위 모듈에 둔다.
+
+### 2. 역할별 자산 격리 (Isolation)
+- **도구/커맨드**: Admiralty 전용 도구는 `admiralty/pi-tools.ts`에서, Fleet 전용 커맨드/도구는 `fleet/` 하위에서만 등록한다.
+- **이벤트**: `pi.on()` 이벤트 구독 소유권을 역할별 `pi-events.ts`로 명확히 분리한다.
+- **키바인딩**: `Alt+G` (Grand Fleet Status) 키바인딩은 Admiralty와 Fleet가 각자의 `status-overlay-keybind.ts`를 통해 독립적으로 등록하며, 팝업 상태(popup guard)도 각자 관리한다.
+
+### 3. 프롬프트 구성 솔기 (Prompt Composition Seam)
+- `prompts.ts`는 Admiralty와 Fleet 모두를 위한 프롬프트 조립 SSOT이다.
+- `buildFleetAcpSystemPrompt()`는 `extensions/fleet/**`를 임포트하지 않고 독자적으로 Fleet ACP 기본 의미와 Grand Fleet Context를 조립하는 **Local Seam** 역할을 한다.
+- Admiralty는 `before_agent_start`에서 시스템 프롬프트를 전체 교체하며, Fleet은 연결 상태일 때만 Context를 append한다.
+
+### 4. globalThis 런타임 버킷 (Runtime Buckets)
+세션 전환 시에도 상태를 유지하기 위해 3가지 전역 버킷을 사용한다:
+- `GRAND_FLEET_STATE_KEY`: 기본 설정 및 공용 상태 (`GrandFleetState`). 역할, 소켓 경로, 연결된 함대 맵 등.
+- `GRAND_FLEET_ADMIRALTY_RUNTIME_KEY`: Admiralty 전용 런타임 핸들 (`AdmiraltyRuntimeState`). IPC 서버, 레지스트리, 디스포저 등.
+- `GRAND_FLEET_FLEET_RUNTIME_KEY`: Fleet 전용 런타임 핸들 (`FleetRuntimeState`). IPC 클라이언트, 하트비트 타이머, 미션 텍스트 버퍼 등.
+
+### 5. 데이터 정화 및 안전 (Sanitization)
+- 오버레이/위젯 등 TUI에 렌더링되는 모든 외부 문자열(함대 ID, 구역, 미션 목표 등)은 `stripControlChars()`로 ANSI/제어문자를 반드시 제거한다.
+- Carrier task 텍스트는 `sanitizeTaskText()`를 통해 민감 토큰 마스킹 및 80자 truncation을 수행한다.
+
+### 6. 슬래시 커맨드 및 커뮤니케이션
+- 슬래시 커맨드 형식: `fleet:grand-fleet:<feature>`.
+- Admiralty는 직접 파일을 수정하거나 쉘 명령을 실행하지 않으며, 오직 함대 파견(`deploy`), 명령 전달(`dispatch`/`broadcast`), 상태 조회(`status`)만 수행한다.
